src/requirements.txt
numpy
scikit-learn
matplotlib
pillow
joblib
